pyproject.toml
[tool.poetry]
name = "monopoly"
version = "0.1.0"
description = ""
authors = ["Plummy <plummybeatsoff@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.7"
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
pydantic-settings = "^2.7.1"
redis = "^5.2.1"
websockets = "^14.2"
arq = "^0.26.3"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
